refactoring

+++ CoE_Challenge/CoE_Challenge/Printer/IPrinter.cs
@@ -4,6 +4,6 @@ using System.Collections.Generic;
 namespace CoE_Challenge.Printer
 {
     public interface IPrinter {
-        void Print(List<IProduct> order, float total, float tax);
+        void Print(List<IProduct> order, decimal total, decimal tax);
     }
 }
