+++ CoE_Challenge/CoE_Challenge/Products/Chips.cs
@@ -3,7 +3,7 @@
 	public class Chips : IProduct
     {
         public string Name { get; } = "Potato chips";
-        public float Price { get; set; } = 10;
+        public decimal Price { get; set; } = 10;
 		public bool IsSnack { get; } = true;
 		public bool IsDiscount { get; } = false;
 	}
