+++ CoE_Challenge/CoE_Challenge/Products/Dip.cs
@@ -3,7 +3,7 @@
 	public class Dip : IProduct
     {
         public string Name { get; } = "Dip";
-        public float Price { get; set; } = 10;
+        public decimal Price { get; set; } = 10;
 		public bool IsSnack { get; } = false;
 		public bool IsDiscount { get; } = false;
     }
