+++ CoE_Challenge/CoE_Challenge/Products/Nachos.cs
@@ -3,7 +3,7 @@
 	public class Nachos : IProduct
     {
         public string Name { get; } = "Nachos";
-        public float Price { get; set; } = 7;
+        public decimal Price { get; set; } = 7;
 		public bool IsSnack { get; } = true;
 		public bool IsDiscount { get; } = false;
     }
