+++ CoE_Challenge/CoE_Challenge/Products/Shampoo.cs
@@ -3,7 +3,7 @@
 	public class Shampoo : IProduct
     {
         public string Name { get; } = "Shampoo";
-        public float Price { get; set; } = 12.95f;
+        public decimal Price { get; set; } = 12.95m;
 		public bool IsSnack { get; } = false;
 		public bool IsDiscount { get; } = false;
     }
