+++ CoE_Challenge/CoE_Challenge/Products/Soap.cs
@@ -3,7 +3,7 @@
 	public class Soap : IProduct
     {
         public string Name { get; } = "Soap";
-        public float Price { get; set; } = 8;
+        public decimal Price { get; set; } = 8;
 		public bool IsSnack { get; } = false;
 		public bool IsDiscount { get; } = false;
     }
