+++ CoE_Challenge/CoE_Challenge/Products/Soda.cs
@@ -3,7 +3,7 @@
 	public class Soda : IProduct
     {
         public string Name { get; } = "Soda (2 lts)";
-        public float Price { get; set; } = 13.5f;
+        public decimal Price { get; set; } = 13.5m;
 		public bool IsSnack { get; } = false;
 		public bool IsDiscount { get; } = false;
     }
