+++ CoE_Challenge/CoE_Challenge/Products/base/IProduct.cs
@@ -4,7 +4,7 @@ namespace CoE_Challenge.Products
 	public interface IProduct
     {
         string Name { get; }
-        float Price { get; set; }
+        decimal Price { get; set; }
         bool IsSnack { get; }
         bool IsDiscount { get; }
     }
